moved item data view to separate component
